decompose main generator file
