+++ src/transformer/transformer.ts
@@ -1,9 +1,11 @@
 import * as ts from 'typescript';
 import { isArray } from 'lodash/fp';
 
-import { generateNodeDescription } from './description-generate';
-import { isTargetExpression } from './utils';
+import { generateNodeDescription } from './type-description-generate/node';
 import { setTypeChecker } from './checker';
+
+const TARGET_CALLERS = ['create', 'createMany', 'build'];
+const TARGET_CLASS_NAME = 'Randomizer';
 
 /**
  * Typescript transformer factory
@@ -15,6 +17,12 @@ export const transformer = (
   setTypeChecker(program.getTypeChecker());
   return ts.visitNode(file, visitNode(context, program.getTypeChecker()));
 };
+
+const isTargetExpression = (target: ts.CallExpression) =>
+  ts.isPropertyAccessExpression(target.expression) &&
+  ts.isIdentifier(target.expression.expression) &&
+  TARGET_CALLERS.includes(target.expression.name.text) &&
+  target.expression.expression.text === TARGET_CLASS_NAME;
 
 /**
  * Typescript AST Node visitor
