+++ src/transformer/type-description-generate/type-params/marge-params-map.ts
@@ -1,0 +1,54 @@
+import { isEmpty } from 'lodash';
+
+import { TypeParamsMap } from '../../../types';
+import { getTypeArguments, getFirstTypeParameter, isArrayType } from '../utils';
+
+export const mergeTypeParamsMap = (
+  argumentsMap: TypeParamsMap = {},
+  nodeTypeArguments: TypeParamsMap = {},
+  typeArgumentsMap: TypeParamsMap = {}
+) => {
+  const newArgumentsMap = isEmpty(argumentsMap)
+    ? isEmpty(nodeTypeArguments) ||
+      Object.keys(nodeTypeArguments).some(k => typeArgumentsMap[k])
+      ? typeArgumentsMap
+      : nodeTypeArguments
+    : argumentsMap;
+
+  Object.keys(newArgumentsMap).forEach(key => {
+    const item = newArgumentsMap[key];
+    const symbol = item.type.getSymbol();
+    if (!symbol) {
+      return;
+    }
+    const found =
+      typeArgumentsMap[symbol.name] || nodeTypeArguments[symbol.name];
+    if (found) {
+      newArgumentsMap[key] = {
+        isArray: item.isArray || found.isArray,
+        type: found.type,
+      };
+      return;
+    }
+    getTypeArguments(item.type).forEach(i => {
+      if (!i.symbol) {
+        return;
+      }
+      const arr = getFirstTypeParameter(i);
+      const symbolName = isArrayType(i)
+        ? arr
+          ? arr.symbol && arr.symbol.name
+          : i.symbol.name
+        : i.symbol.name;
+      if (!symbolName) {
+        return;
+      }
+      newArgumentsMap[symbolName] =
+        typeArgumentsMap[symbolName] ||
+        nodeTypeArguments[symbolName] ||
+        newArgumentsMap[symbolName];
+    });
+  });
+
+  return newArgumentsMap;
+};
